requirements.txt
torch
pdf2image
pypdf
tqdm
pillow
colpali-engine
pyvespa
requests
numpy
python-dotenv



